+++ src/types/string.rs
@@ -24,7 +24,12 @@ pyobject_native_type!(PyString, ffi::PyUnicode_Type, ffi::PyUnicode_Check);
 #[repr(transparent)]
 pub struct PyBytes(PyObject);
 
-pyobject_native_type!(PyBytes, ffi::PyBytes_Type, ffi::PyBytes_Check);
+pyobject_native_type!(
+    PyBytes,
+    ffi::PyBytes_Type,
+    Some("builtins"),
+    ffi::PyBytes_Check
+);
 
 impl PyString {
     /// Creates a new Python string object.