+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+ * `module` argument to `pyclass` macro. [#499](https://github.com/PyO3/pyo3/pull/499)
+
+
 ## [0.7.0] - 2018-05-26
 
 ### Added